Force UTF-8 Encoding when reading the NuSpec
Hi! First of all, thanks to the AWESOME plugin. Good work!
I found one minor issue when my Nuget description has special characters, even if the encoding is defined on the XML tag like this:
<?xml version="1.0" encoding="UTF-8"?>
So, if you force UTF-8 it via the Get-Content we would be on safe side.