WIQL pull from workItemRelations for WorkItemLinks
The WIQL flow was not handling the alternate return object for when the wiqlFromTarget was set to WorkItemLinks resulting in no errors but no results ever processed even when the query had results. Since various linkage types are returned I decided to just pull all and dedupe them.

Not positive if target?.id is the correct source for all flows though.